Move CategoricalChartState to its own file
