+++ src/chart/generateCategoricalChart.tsx
@@ -87,7 +87,8 @@ import { AccessibilityManager } from './AccessibilityManager';
 import { isDomainSpecifiedByUser } from '../util/isDomainSpecifiedByUser';
 import { getActiveShapeIndexForTooltip, isFunnel, isPie, isScatter } from '../util/ActiveShapeUtils';
 import { Cursor } from '../component/Cursor';
-import { ChartLayoutContext } from '../context/chartLayoutContext';
+import { ChartLayoutContextProvider } from '../context/chartLayoutContext';
+import { AxisMap, CategoricalChartState } from './types';
 
 export interface MousePointer {
   pageX: number;
@@ -792,83 +793,9 @@ const calculateOffset = (
   };
 };
 
-type AxisMap = {
-  [axisId: string]: BaseAxisProps;
-};
-
 type AxisMapMap = {
   [axisMapId: string]: AxisMap;
 };
-
-export interface CategoricalChartState {
-  chartX?: number;
-
-  chartY?: number;
-
-  dataStartIndex?: number;
-
-  dataEndIndex?: number;
-
-  activeTooltipIndex?: number;
-
-  isTooltipActive?: boolean;
-
-  updateId?: number;
-
-  xAxisMap?: AxisMap;
-
-  yAxisMap?: AxisMap;
-
-  zAxisMap?: AxisMap;
-
-  orderedTooltipTicks?: any;
-
-  tooltipAxis?: BaseAxisProps;
-
-  tooltipTicks?: TickItem[];
-
-  graphicalItems?: ReadonlyArray<ReactElement>;
-
-  activeCoordinate?: ChartCoordinate;
-
-  offset?: ChartOffset;
-
-  angleAxisMap?: any;
-
-  radiusAxisMap?: any;
-
-  formattedGraphicalItems?: any;
-
-  /** active tooltip payload */
-  activePayload?: any[];
-
-  tooltipAxisBandSize?: number;
-
-  /** active item */
-  activeItem?: any;
-
-  /** Active label of data */
-  activeLabel?: string;
-
-  activeIndex?: number;
-
-  xValue?: number;
-
-  yValue?: number;
-
-  legendBBox?: DOMRect | null;
-
-  prevDataKey?: DataKey<any>;
-  prevData?: any[];
-  prevWidth?: number;
-  prevHeight?: number;
-  prevLayout?: LayoutType;
-  prevStackOffset?: StackOffsetType;
-  prevMargin?: Margin;
-  prevChildren?: any;
-
-  stackGroups?: AxisStackGroups;
-}
 
 export type CategoricalChartFunc = (nextState: CategoricalChartState, event: any) => void;
 
@@ -2303,7 +2230,6 @@ export const generateCategoricalChart = ({
       }
 
       const { children, className, width, height, style, compact, title, desc, ...others } = this.props;
-      const { xAxisMap, yAxisMap, offset } = this.state;
       const attrs = filterProps(others);
 
       // The "compact" mode is mainly used as the panorama within Brush
@@ -2335,19 +2261,7 @@ export const generateCategoricalChart = ({
 
       const events = this.parseEventsOfWrapper();
       return (
-        <ChartLayoutContext.Provider
-          value={{
-            xAxisMap,
-            yAxisMap,
-            viewBox: {
-              x: offset.left,
-              y: offset.top,
-              width: offset.width,
-              height: offset.height,
-            },
-            clipPathId: this.clipPathId,
-          }}
-        >
+        <ChartLayoutContextProvider value={this.state}>
           <div
             className={clsx('recharts-wrapper', className)}
             style={{ position: 'relative', cursor: 'default', width, height, ...style }}
@@ -2364,7 +2278,7 @@ export const generateCategoricalChart = ({
             {this.renderLegend()}
             {this.renderTooltip()}
           </div>
-        </ChartLayoutContext.Provider>
+        </ChartLayoutContextProvider>
       );
     }
   };
